backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3
sqlmodel==0.0.21
psycopg2-binary==2.9.10
numpy>=1.24.0
faiss-cpu>=1.7.4
pytest>=7.0.0
